중간 체크인

+++ ThisApp.cpp
@@ -36,7 +36,8 @@ public:
 		render.reset(new RenderProcedure(hWnd, global->d3d11.get(), global->rts.get()));
 
 		arcBall.reset(IEulerControl::Create(
-			210 / 180.0f * M_PI, 20 / 180.0f * M_PI));
+			210 / 180.0f * (float)M_PI, 
+			20 / 180.0f * (float)M_PI));
 	}
 
 	//--------------------------------------------------------------------------
@@ -106,7 +107,7 @@ public:
 					it != global->logger->entry.end(); ++it)
 				{
 					global->textRenderer->Enqueue(TextToRender(
-						XMFLOAT2(50, 50 + y * 20),
+						XMFLOAT2(50.0f, 50.0f + y * 20),
 						*it,
 						XMFLOAT4(1, 0, 0, 1)));
 
@@ -116,8 +117,8 @@ public:
 				global->textRenderer->Draw(
 					sceneDesc.worldViewProj,
 					XMFLOAT2(
-						global->rts->GetCurrent()->GetWidth(),
-						global->rts->GetCurrent()->GetHeight()));
+						(float)global->rts->GetCurrent()->GetWidth(),
+						(float)global->rts->GetCurrent()->GetHeight()));
 			}
 
 			render->End();
